[#4] Updated based on PR comments
