--- app/views/accounts/add_nation/new.html.haml
@@ -1,1 +1,0 @@
-= render 'modal'
