+++ app/views/accounts/nations/_modal.html.haml
@@ -5,6 +5,6 @@
       %button.close{"aria-label" => "Close", "data-dismiss" => "modal", :type => "button"}
         %span{"aria-hidden" => "true"} ×
     .modal-body
-      = bootstrap_form_for @nation, url: accounts_add_nation_index_path(account_id: params[:account_id]) do |f|
+      = bootstrap_form_for @nation, url: accounts_nations_path(account_id: params[:account_id]) do |f|
         = f.text_field :slug
         = f.submit 'Add Nation', class: 'btn btn-primary btn-sm add-nation-submit'
