+++ app/views/accounts/raisely_campaigns/_new_modal.html.haml
@@ -5,7 +5,7 @@
       %button.close{"aria-label" => "Close", "data-dismiss" => "modal", :type => "button"}
         %span{"aria-hidden" => "true"} ×
     .modal-body
-      = bootstrap_form_for @campaign, url: accounts_add_campaign_index_path(account_id: params[:account_id]) do |f|
+      = bootstrap_form_for @campaign, url: accounts_raisely_campaigns_path(account_id: params[:account_id]) do |f|
         = f.text_field :campaign_uuid
         = f.text_field :api_key
         = f.submit 'Add Campaign', class: 'btn btn-primary btn-sm add-campaign-submit'
