+++ app/views/accounts/raisely_campaigns/edit.html.haml
@@ -1,1 +1,1 @@
-= render 'modal'
+= render 'update_modal'
