+++ app/views/accounts/raisely_campaigns/new.html.haml
@@ -1,0 +1,1 @@
+= render 'new_modal'
