+++ config/environments/development.rb
@@ -5,7 +5,9 @@ Rails.application.configure do
   # every request. This slows down response time but is perfect for development
   # since you don't have to restart the web server when you make code changes.
   config.cache_classes = false
-  config.hosts << '473e76c3cf0b.ngrok.io'
+
+  # allow any hosts
+  config.hosts = nil
 
   # Do not eager load code on boot.
   config.eager_load = false
